[FIX] rename directories and fields
