+++ server/src/main/java/mx/edu/utez/server/config/InitialConfig.java
@@ -6,8 +6,8 @@ import mx.edu.utez.server.kernel.Roles;
 import mx.edu.utez.server.kernel.Shifts;
 import mx.edu.utez.server.kernel.StatusType;
 import mx.edu.utez.server.kernel.Statuses;
-import mx.edu.utez.server.modules.appointmentType.service.AppointmentTypeService;
-import mx.edu.utez.server.modules.cancellationReason.service.CancellationReasonService;
+import mx.edu.utez.server.modules.appointment_type.service.AppointmentTypeService;
+import mx.edu.utez.server.modules.cancellation_reason.service.CancellationReasonService;
 import mx.edu.utez.server.modules.gender.service.GenderService;
 import mx.edu.utez.server.modules.role.service.RoleService;
 import mx.edu.utez.server.modules.shift.service.ShiftService;
