+++ server/src/main/java/mx/edu/utez/server/modules/appointment/model/Appointment.java
@@ -20,8 +20,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import mx.edu.utez.server.modules.appointmentType.model.AppointmentType;
-import mx.edu.utez.server.modules.cancellationReason.module.CancellationReason;
+import mx.edu.utez.server.modules.appointment_type.model.AppointmentType;
+import mx.edu.utez.server.modules.cancellation_reason.module.CancellationReason;
 import mx.edu.utez.server.modules.doctor.model.Doctor;
 import mx.edu.utez.server.modules.patient.model.Patient;
 import mx.edu.utez.server.modules.record.model.Record;
