--- server/src/main/java/mx/edu/utez/server/modules/appointmentType/controller/dto/AppointmentTypeDto.java
@@ -1,7 +1,0 @@
-package mx.edu.utez.server.modules.appointmentType.controller.dto;
-
-import lombok.Data;
-
-@Data
-public class AppointmentTypeDto {
-}
