--- server/src/main/java/mx/edu/utez/server/modules/appointmentType/model/AppointmentType.java
@@ -1,55 +1,0 @@
-package mx.edu.utez.server.modules.appointmentType.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import mx.edu.utez.server.modules.appointment.model.Appointment;
-import mx.edu.utez.server.modules.status.model.Status;
-
-import java.util.Set;
-
-@Entity
-@Table(name = "appointment_types")
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class AppointmentType {
-    public AppointmentType(String name, String description, Status status) {
-        this.name = name;
-        this.description = description;
-        this.status = status;
-    }
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(columnDefinition = "VARCHAR(45)", nullable = false, unique = true)
-    private String name;
-
-    @Column(columnDefinition = "VARCHAR(200)")
-    private String description;
-
-    // Relationships <-
-    @ManyToOne
-    @JoinColumn(name = "status_id", referencedColumnName = "id",
-            nullable = false)
-    private Status status;
-
-    // Relationships ->
-    @OneToMany(mappedBy = "appointmentType")
-    @JsonIgnore
-    private Set<Appointment> appointments;
-}
