+++ server/src/main/java/mx/edu/utez/server/modules/appointment_type/controller/dto/AppointmentTypeDto.java
@@ -1,0 +1,7 @@
+package mx.edu.utez.server.modules.appointment_type.controller.dto;
+
+import lombok.Data;
+
+@Data
+public class AppointmentTypeDto {
+}
