+++ server/src/main/java/mx/edu/utez/server/modules/status/model/Status.java
@@ -17,9 +17,9 @@ import lombok.Setter;
 import mx.edu.utez.server.kernel.StatusType;
 import mx.edu.utez.server.kernel.Statuses;
 import mx.edu.utez.server.modules.appointment.model.Appointment;
-import mx.edu.utez.server.modules.appointmentType.model.AppointmentType;
+import mx.edu.utez.server.modules.appointment_type.model.AppointmentType;
 import mx.edu.utez.server.modules.area.model.Area;
-import mx.edu.utez.server.modules.cancellationReason.module.CancellationReason;
+import mx.edu.utez.server.modules.cancellation_reason.module.CancellationReason;
 import mx.edu.utez.server.modules.doctor.model.Doctor;
 import mx.edu.utez.server.modules.gender.model.Gender;
 import mx.edu.utez.server.modules.patient.model.Patient;
