--- server/src/main/java/mx/edu/utez/server/modules/verificationCode/controller/VerificationCodeController.java
@@ -1,15 +1,0 @@
-package mx.edu.utez.server.modules.verificationCode.controller;
-
-import lombok.RequiredArgsConstructor;
-import mx.edu.utez.server.modules.verificationCode.service.VerificationCodeService;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping(value = "/api/verification-code")
-@CrossOrigin(origins = {"*"})
-@RequiredArgsConstructor
-public class VerificationCodeController {
-    private final VerificationCodeService verificationCodeService;
-}
