--- server/src/main/java/mx/edu/utez/server/modules/verificationCode/controller/dto/VerificationCodeDto.java
@@ -1,7 +1,0 @@
-package mx.edu.utez.server.modules.verificationCode.controller.dto;
-
-import lombok.Data;
-
-@Data
-public class VerificationCodeDto {
-}
