+++ server/src/main/java/mx/edu/utez/server/modules/verification_code/controller/dto/VerificationCodeDto.java
@@ -1,0 +1,7 @@
+package mx.edu.utez.server.modules.verification_code.controller.dto;
+
+import lombok.Data;
+
+@Data
+public class VerificationCodeDto {
+}
