+++ server/src/main/java/mx/edu/utez/server/modules/verification_code/model/IVerificationCodeRepository.java
@@ -1,0 +1,8 @@
+package mx.edu.utez.server.modules.verification_code.model;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IVerificationCodeRepository extends JpaRepository<VerificationCode, Long> {
+}
