+++ server/src/main/java/mx/edu/utez/server/modules/verification_code/service/VerificationCodeService.java
@@ -1,0 +1,13 @@
+package mx.edu.utez.server.modules.verification_code.service;
+
+import lombok.RequiredArgsConstructor;
+import mx.edu.utez.server.modules.verification_code.model.IVerificationCodeRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class VerificationCodeService {
+    private final IVerificationCodeRepository iVerificationCodeRepository;
+}
