--- server/src/main/java/mx/edu/utez/server/utils/dtoGroups/GetList.java
@@ -1,4 +1,0 @@
-package mx.edu.utez.server.utils.dtoGroups;
-
-public interface GetList {
-}
