--- server/src/main/java/mx/edu/utez/server/utils/dtoGroups/Update.java
@@ -1,4 +1,0 @@
-package mx.edu.utez.server.utils.dtoGroups;
-
-public interface Update {
-}
