+++ server/src/main/java/mx/edu/utez/server/utils/dto_groups/Update.java
@@ -1,0 +1,4 @@
+package mx.edu.utez.server.utils.dto_groups;
+
+public interface Update {
+}
